reprogrammation rdv

+++ lib/components/rendez_vous_table.dart
@@ -2,6 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_svg/svg.dart';
 import 'package:intl/intl.dart';
 import 'package:medstory/components/box.dart';
+import 'package:medstory/components/rdv_reprog.dart';
+import 'package:medstory/components/time_slot_selector.dart';
 import 'package:medstory/constantes.dart';
 import 'package:medstory/models/medecin.dart';
 import 'package:medstory/models/my_data.dart';
@@ -126,23 +128,29 @@ class _RendezVousTableState extends State<RendezVousTable> {
                       height: 25,
                       width: 25,
                     ),
-                    onPressed: () async {
-                      // Action pour supprimer le médecin
-                      context.showLoader();
-                      final rendezVousService = RendezVousService();
-                      await rendezVousService
-                          .deleteRendezVous(rendezVous.id)
-                          .then((value) {
-                        context
-                            .read<MyData>()
-                            .fetchRendezVous(); //TODO: Remplacer la logique (fetch only for current user)
-                        context.hideLoader();
-                        context.showSuccess(
-                            "Le rendez-vous a été supprimé avec succès.");
-                      }).catchError((onError) {
-                        context.hideLoader();
-                        context.showError(onError.toString());
-                      });
+                    onPressed: () {
+                      context.showConfirmation(
+                          title: "Attention !!",
+                          message:
+                              "Cette action supprime le rendez-vous.\nÊtes-vous sûr(e) de continuer ?",
+                          onConfirm: () async {
+                            // Action pour supprimer le médecin
+                            context.showLoader();
+                            final rendezVousService = RendezVousService();
+                            await rendezVousService
+                                .deleteRendezVous(rendezVous.id)
+                                .then((value) {
+                              context
+                                  .read<MyData>()
+                                  .fetchRendezVous(); //TODO: Remplacer la logique (fetch only for current user)
+                              context.hideLoader();
+                              context.showSuccess(
+                                  "Le rendez-vous a été supprimé avec succès.");
+                            }).catchError((onError) {
+                              context.hideLoader();
+                              context.showError(onError.toString());
+                            });
+                          });
                     },
                   ),
                 ],
@@ -151,8 +159,11 @@ class _RendezVousTableState extends State<RendezVousTable> {
                   '${rendezVous.patient.prenom} ${rendezVous.patient.nom}')),
               DataCell(Text(
                   '${rendezVous.medecin.prenom} ${rendezVous.medecin.nom}')),
-              DataCell(Text(
-                  '${rendezVous.date.day}/${rendezVous.date.month}/${rendezVous.date.year}')),
+              DataCell(
+                Text(
+                  DateFormat('dd-MM-yyyy').format(rendezVous.date),
+                ),
+              ),
               DataCell(Text(rendezVous.heure)),
               DataCell(Text(rendezVous.motif)),
               DataCell(Text(rendezVous.statut.libelle)),
@@ -167,302 +178,16 @@ class _RendezVousTableState extends State<RendezVousTable> {
     return showDialog(
         context: contexte,
         builder: (context) {
-          return FractionallySizedBox(
-            heightFactor: 0.85,
-            widthFactor: 0.8,
-            child: Box(
-                child: Column(
-              mainAxisSize: MainAxisSize.min,
-              children: [
-                const Text("Nouvelle programmation",
-                    style: TextStyle(fontSize: 18)),
-                const SizedBox(height: 16),
-                Expanded(
-                  child: SingleChildScrollView(
-                    child: Column(
-                      children: [
-                        Text(
-                          "Medecin: Dr ${rdv.medecin.nom} ${rdv.medecin.prenom}",
-                          style:
-                              Theme.of(context).textTheme.bodyLarge!.copyWith(
-                                    color: Colors.black,
-                                    // fontSize: 18,
-                                  ),
-                        ),
-                        Text(
-                          "Patient: ${rdv.patient.nom} ${rdv.patient.prenom}",
-                          style:
-                              Theme.of(context).textTheme.bodyLarge!.copyWith(
-                                    color: Colors.black,
-                                    // fontSize: 18,
-                                  ),
-                        ),
-                        Text(
-                          "Motif: ${rdv.motif}",
-                          style:
-                              Theme.of(context).textTheme.bodyLarge!.copyWith(
-                                    color: Colors.black,
-                                    // fontSize: 18,
-                                  ),
-                        ),
-
-                        // Date de rendez-vous
-                        Row(
-                          children: [
-                            Column(
-                              children: [
-                                const Text("Date de rendez-vous"),
-                                const SizedBox(height: 8),
-                                ElevatedButton(
-                                  onPressed: () {
-                                    _selectDate(
-                                      context,
-                                      rdv.medecin,
-                                    );
-                                    print(rdv.medecin.joursIntervention);
-                                  },
-                                  style: ElevatedButton.styleFrom(
-                                    backgroundColor: primaryColor,
-                                    shape: RoundedRectangleBorder(
-                                      borderRadius: BorderRadius.circular(5),
-                                    ),
-                                  ),
-                                  child: Text(
-                                    _selectedDate != null
-                                        ? DateFormat('dd/MM/yyyy')
-                                            .format(_selectedDate!)
-                                        : "Sélectionner une date",
-                                    style: const TextStyle(
-                                      color: Colors.white,
-                                    ),
-                                  ),
-                                ),
-                              ],
-                            ),
-                            const SizedBox(width: 16),
-                            // Heure
-                            Column(
-                              children: [
-                                const Text("Heure de rendez-vous"),
-                                const SizedBox(height: 8),
-                                ElevatedButton(
-                                  onPressed: () => _selectTime(
-                                    contexte,
-                                    rdv.medecin,
-                                  ),
-                                  style: ElevatedButton.styleFrom(
-                                    backgroundColor: tertiaryColor,
-                                    shape: RoundedRectangleBorder(
-                                      borderRadius: BorderRadius.circular(5),
-                                    ),
-                                  ),
-                                  child: Text(
-                                    _selectedTime != null
-                                        ? _selectedTime!.format(context)
-                                        : "Sélectionner une heure",
-                                    style: const TextStyle(
-                                      color: Colors.white,
-                                    ),
-                                  ),
-                                ),
-                              ],
-                            ),
-                          ],
-                        ),
-                        const SizedBox(
-                          height: 20,
-                        ),
-                        Row(
-                          children: [
-                            Expanded(
-                              child: ElevatedButton(
-                                onPressed: () async {
-                                  // context.showLoader();
-                                  // final rendezVousService = RendezVousService();
-
-                                  // RendezVous rdv = RendezVous(
-                                  //     id: 0,
-                                  //     motif: motifController.text,
-                                  //     date: _selectedDate!,
-                                  //     heure: heureRdv!,
-                                  //     statut: Statut(id: 0, libelle: "en attente"),
-                                  //     medecin: selectedMedecin!,
-                                  //     patient: selectedPatient!);
-
-                                  // // TODO: à des fin de vidéo utiliser cette logige
-                                  // Future.delayed(const Duration(milliseconds: 300), () {
-                                  //   context.read<MyData>().addRendezVous(rdv);
-                                  //   context.hideLoader();
-                                  //   context.showSuccess(
-                                  //       "Le rendez-vous a été ajouté avec succès.");
-                                  //   widget.changeView();
-                                  // });
-                                },
-                                style: ElevatedButton.styleFrom(
-                                  backgroundColor: primaryColor,
-                                  shape: RoundedRectangleBorder(
-                                    borderRadius: BorderRadius.circular(5),
-                                  ),
-                                ),
-                                child: const Text(
-                                  "Confirmer le rendez-vous",
-                                  style: TextStyle(
-                                    color: Colors.white,
-                                  ),
-                                ),
-                              ),
-                            ),
-                            const SizedBox(
-                              width: 12,
-                            ),
-                            Expanded(
-                              child: ElevatedButton(
-                                onPressed: () {}, // TODO: Logique cancel ici.
-                                style: ElevatedButton.styleFrom(
-                                  backgroundColor: Colors.black,
-                                  shape: RoundedRectangleBorder(
-                                    borderRadius: BorderRadius.circular(5),
-                                  ),
-                                ),
-                                child: const Text(
-                                  "Annuler",
-                                  style: TextStyle(
-                                    color: Colors.white,
-                                  ),
-                                ),
-                              ),
-                            ),
-                          ],
-                        ),
-                      ],
-                    ),
-                  ),
-                ),
-              ],
-            )),
+          return Material(
+            type: MaterialType.transparency,
+            child: FractionallySizedBox(
+              heightFactor: 0.85,
+              widthFactor: 0.8,
+              child: RdvReprog(
+                rdv: rdv,
+              ),
+            ),
           );
         });
-  }
-
-  void _selectDate(BuildContext context, Medecin? selectedMedecin) async {
-    if (selectedMedecin == null) {
-      ScaffoldMessenger.of(context).showSnackBar(
-        const SnackBar(
-          content: Text("Veuillez d'abord sélectionner un médecin."),
-        ),
-      );
-      return;
-    }
-
-    DateTime initialDate = getNextAvailableDate(selectedMedecin);
-
-    DateTime today = initialDate;
-    DateTime? selectedDate = await showDatePicker(
-      context: context,
-      initialDate: today,
-      firstDate: today,
-      lastDate: today.add(const Duration(days: 365)),
-      selectableDayPredicate: (DateTime day) {
-        String dayOfWeek = DateFormat('EEEE').format(day).toUpperCase();
-        return selectedMedecin.joursIntervention
-            .map((e) => e['jour'] as String)
-            .contains(dayOfWeek);
-      },
-    );
-
-    if (selectedDate != null) {
-      setState(() {
-        _selectedDate = selectedDate;
-      });
-    }
-  }
-
-  void _selectTime(BuildContext context, Medecin? selectedMedecin) async {
-    if (_selectedDate == null) {
-      ScaffoldMessenger.of(context).showSnackBar(
-        const SnackBar(
-          content: Text("Veuillez d'abord sélectionner une date."),
-        ),
-      );
-      return;
-    }
-
-    String dayOfWeek = DateFormat('EEEE').format(_selectedDate!).toUpperCase();
-    Map<String, String> heuresIntervention =
-        getHeuresPourJour(dayOfWeek, selectedMedecin);
-
-    List<String> debutSplit = heuresIntervention['heureDebut']!.split(":");
-    List<String> finSplit = heuresIntervention['heureFin']!.split(":");
-    TimeOfDay heureDebut = TimeOfDay(
-        hour: int.parse(debutSplit[0]), minute: int.parse(debutSplit[1]));
-    TimeOfDay heureFin =
-        TimeOfDay(hour: int.parse(finSplit[0]), minute: int.parse(finSplit[1]));
-
-    TimeOfDay? selectedTime = await showTimePicker(
-      context: context,
-      initialTime: heureDebut,
-      builder: (context, child) {
-        return MediaQuery(
-          data: MediaQuery.of(context).copyWith(alwaysUse24HourFormat: true),
-          child: child!,
-        );
-      },
-    );
-
-    if (selectedTime != null &&
-        (selectedTime.hour >= heureDebut.hour &&
-            selectedTime.hour <= heureFin.hour)) {
-      setState(() {
-        _selectedTime = selectedTime;
-        heureRdv =
-            '${selectedTime.hour.toString().padLeft(2, '0')}:${selectedTime.minute.toString().padLeft(2, '0')}';
-      });
-    } else {
-      ScaffoldMessenger.of(context).showSnackBar(
-        const SnackBar(
-          content: Text(
-            "Veuillez sélectionner une heure dans l'intervalle disponible.",
-          ),
-        ),
-      );
-    }
-  }
-
-  Map<String, String> getHeuresPourJour(String jour, Medecin? selectedMedecin) {
-    if (selectedMedecin == null) {
-      return {'heureDebut': '00:00', 'heureFin': '23:59'};
-    }
-
-    Map<String, dynamic> intervention =
-        selectedMedecin.joursIntervention.firstWhere(
-      (e) => e['jour'] == jour,
-      orElse: () => {},
-    );
-
-    print("le jour ::: $intervention");
-
-    // return {
-    //   'heureDebut': '00:00',
-    //   'heureFin': '23:59',
-    // };
-    return {
-      'heureDebut': intervention['heureDebut'] ?? '00:00',
-      'heureFin': intervention['heureFin'] ?? '23:59',
-    };
-  }
-
-  DateTime getNextAvailableDate(Medecin medecin) {
-    DateTime today = DateTime.now();
-    for (int i = 0; i < 7; i++) {
-      DateTime dateToCheck = today.add(Duration(days: i));
-      String dayOfWeek = DateFormat('EEEE').format(dateToCheck).toUpperCase();
-      bool isInterventionDay = medecin.joursIntervention
-          .map((e) => e['jour'] as String)
-          .contains(dayOfWeek);
-      if (isInterventionDay) {
-        return dateToCheck;
-      }
-    }
-    return today; // Retourne la date d'aujourd'hui si aucun jour d'intervention n'est trouvé
   }
 }
